requirements.txt
aeventkit==2.1.0
alpaca-py==0.43.2
altair==6.0.0
annotated-types==0.7.0
attrs==26.1.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
gitdb==4.0.12
GitPython==3.1.46
ib_async==2.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
llvmlite==0.44.0
MarkupSafe==3.0.3
msgpack==1.1.2
narwhals==2.18.1
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pandas-ta==0.4.71b0
pillow==12.1.1
plotly==6.6.0
protobuf==6.33.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
referencing==0.37.0
requests==2.33.0
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sseclient-py==1.9.0
streamlit==1.55.0
tenacity==9.1.4
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==16.0
